Moved paths and PathManager to utils.
