+++ fast_pedago/gui/components/app_layout.py
@@ -8,8 +8,7 @@ from .input_widgets import (
     GitLinksButton,
     ClearAllButton,
 )
-from fast_pedago.processes import PathManager
-from fast_pedago.utils.functions import _image_from_path
+from fast_pedago.utils import _image_from_path, PathManager
 
 
 # Components sizes
