--- fast_pedago/notebook/workdir/outputs/aircraft_mda_flight_points.csv
@@ -1,14 +1,0 @@
-,time [s],altitude [m],isa_offset,ground_distance [m],mass [kg],consumed_fuel [kg],true_airspeed [m/s],equivalent_airspeed [m/s],mach [-],engine_setting,CL [-],CD [-],lift [N],drag [N],thrust [N],thrust_rate [-],thrust_is_regulated,sfc [kg/N/s],slope_angle [rad],acceleration [m/s**2],alpha [rad],slope_angle_derivative [rad/s],name
-0,0.0,0.0,0.0,0.0,95482.34855488496,0.0,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-1,60.0,0.0,0.0,0.0,95446.70952006472,35.63903482024594,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-2,120.0,0.0,0.0,0.0,95411.07048524448,71.27806964049188,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-3,180.0,0.0,0.0,0.0,95375.43145042424,106.91710446073782,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-4,240.0,0.0,0.0,0.0,95339.792415604,142.55613928098376,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-5,300.0,0.0,0.0,0.0,95304.15338078375,178.1951741012297,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-6,300.0,10.668000000000001,0.0,0.0,95221.75338078376,260.595174101225,79.3,79.2593976814894,0.23306096703307572,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:takeoff
-7,300.0,10.668000000000001,0.0,0.0,95221.75338078376,260.595174101225,79.3,79.2593976814894,0.23306096703307572,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:takeoff
-8,300.0,10668.0,0.0,250000.0,92365.10077936025,3117.2477755247382,222.40239239916536,123.80352173708312,0.75,2,0.6748193640504069,0.042418101050184176,,56936.6971027461,63502.78413057599,0.3,True,9.353667273066347e-06,0.0,0.0,0.0,,sizing:main_route:climb
-9,23033.56840031445,10668.0,0.0,5306000.0,74201.32770276807,21281.02085211691,222.40239239916536,123.80352173708312,0.75,3,0.5421148501931469,0.03213392151175596,,43132.51443466779,43132.51443466779,0.7873320280135546,True,1.6338180363997536e-05,0.0,0.0,0.0,,sizing:main_route:cruise
-10,23033.56840031445,0.0,0.0,5556000.0,72717.30114871271,22765.047406172278,0.0,0.0,0.0,3,0.5421148501931469,0.03213392151175596,,43132.51443466779,43132.51443466779,0.7873320280135546,True,1.6338180363997536e-05,0.0,0.0,0.0,,sizing:main_route:descent
-11,23033.56840031445,0.0,0.0,5556000.0,72717.30114871271,22765.047406172278,0.0,0.0,0.0,3,0.5421148501931469,0.03213392151175596,,43132.51443466779,43132.51443466779,0.7873320280135546,True,1.6338180363997536e-05,0.0,0.0,0.0,,sizing:global_reserve
-12,23033.56840031445,0.0,0.0,5556000.0,68601.22749878556,26881.121056099422,0.0,0.0,0.0,3,0.5421148501931469,0.03213392151175596,,43132.51443466779,43132.51443466779,0.7873320280135546,True,1.6338180363997536e-05,0.0,0.0,0.0,,sizing:global_reserve
