--- fast_pedago/notebook/workdir/outputs/flight_points.csv
@@ -1,14 +1,0 @@
-,time [s],altitude [m],isa_offset,ground_distance [m],mass [kg],consumed_fuel [kg],true_airspeed [m/s],equivalent_airspeed [m/s],mach [-],engine_setting,CL [-],CD [-],lift [N],drag [N],thrust [N],thrust_rate [-],thrust_is_regulated,sfc [kg/N/s],slope_angle [rad],acceleration [m/s**2],alpha [rad],slope_angle_derivative [rad/s],name
-0,0.0,0.0,0.0,0.0,97139.4953761398,0.0,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-1,60.0,0.0,0.0,0.0,97103.85634131957,35.63903482024594,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-2,120.0,0.0,0.0,0.0,97068.21730649933,71.27806964049188,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-3,180.0,0.0,0.0,0.0,97032.57827167909,106.91710446073782,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-4,240.0,0.0,0.0,0.0,96996.93923685885,142.55613928098376,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-5,300.0,0.0,0.0,0.0,96961.3002020386,178.1951741012297,0.0,0.0,0.0,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:taxi_out
-6,300.0,10.668000000000001,0.0,0.0,96878.90020203861,260.595174101225,79.3,79.2593976814894,0.23306096703307572,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:takeoff
-7,300.0,10.668000000000001,0.0,0.0,96878.90020203861,260.595174101225,79.3,79.2593976814894,0.23306096703307572,2,0.0,0.0,,0.0,63502.78413057599,0.3,False,9.353667273066347e-06,0.0,0.0,0.0,,sizing:takeoff
-8,300.0,10668.0,0.0,250000.0,94456.92769698764,2682.5676791521983,222.40239239916536,123.80352173708312,0.75,2,0.652906887698577,0.034527131506401576,,48985.07078965964,63502.78413057599,0.3,True,9.353667273066347e-06,0.0,0.0,0.0,,sizing:main_route:climb
-9,23033.56840031445,10668.0,0.0,5306000.0,78241.83929233927,18897.656083800568,222.40239239916536,123.80352173708312,0.75,3,0.5408246597226836,0.028235323302333144,,40058.62202822373,40058.62202822373,0.7312218296172409,True,1.636835648211422e-05,0.0,0.0,0.0,,sizing:main_route:cruise
-10,23033.56840031445,0.0,0.0,5556000.0,77694.1464172929,19445.34895884694,0.0,0.0,0.0,3,0.5408246597226836,0.028235323302333144,,40058.62202822373,40058.62202822373,0.7312218296172409,True,1.636835648211422e-05,0.0,0.0,0.0,,sizing:main_route:descent
-11,23033.56840031445,0.0,0.0,5556000.0,77694.1464172929,19445.34895884694,0.0,0.0,0.0,3,0.5408246597226836,0.028235323302333144,,40058.62202822373,40058.62202822373,0.7312218296172409,True,1.636835648211422e-05,0.0,0.0,0.0,,sizing:global_reserve
-12,23033.56840031445,0.0,0.0,5556000.0,73643.74067989849,23495.754696241347,0.0,0.0,0.0,3,0.5408246597226836,0.028235323302333144,,40058.62202822373,40058.62202822373,0.7312218296172409,True,1.636835648211422e-05,0.0,0.0,0.0,,sizing:global_reserve
