--- fast_pedago/objects/__init__.py
@@ -1,22 +1,0 @@
-from .paths import (
-    SOURCE_FILE_SUFFIX,
-    INPUT_FILE_SUFFIX,
-    OUTPUT_FILE_SUFFIX,
-    FLIGHT_DATA_FILE_SUFFIX,
-    RECORDER_FILE_SUFFIX,
-    MDA_FILE_SUFFIX,
-    MDO_FILE_SUFFIX,
-    MDA,
-    MDO,
-    SEPARATOR,
-    WORK_DIRECTORY,
-    DATA_DIRECTORY,
-    INPUTS_DIRECTORY,
-    OUTPUTS_DIRECTORY,
-    RESOURCES_DIRECTORY,
-    TUTORIAL_DIRECTORY,
-    MDA_CONFIGURATION_FILE,
-    MDO_CONFIGURATION_FILE,
-    REFERENCE_AIRCRAFT,
-    DEFAULT_PROCESS_NAME,
-)
